Fix Delta schema mismatch by ensuring proper integer types
Changes:
- Explicitly cast date_key to int (not Int64) to avoid NaN issues
- Cast location_key to Int64 to handle missing values properly
- Cast lob_key to int for consistency

Resolves: DELTA_FAILED_TO_MERGE_FIELDS error when publishing to Databricks

+++ create_fact_occupancy.py
@@ -194,9 +194,11 @@ def create_fact_occupancy():
     # Fill missing attendance with 0
     fact_table['attendance_count'] = fact_table['attendance_count'].fillna(0).astype(int)
 
-    # Ensure location_key and lob_key are integers (not floats)
+    # Ensure date_key, location_key and lob_key are proper integers (not floats)
+    # Use Int64 for keys that might have NaN values
+    fact_table['date_key'] = fact_table['date_key'].astype(int)
     fact_table['location_key'] = fact_table['location_key'].astype('Int64')
-    fact_table['lob_key'] = fact_table['lob_key'].astype('Int64')
+    fact_table['lob_key'] = fact_table['lob_key'].astype(int)
 
     print(f"Fact table now has {len(fact_table)} rows with complete coverage")
